+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - refactored protocols and scenario management
 - refactored formatting options for better parsing and granularity
+- changed output to uniform between single- and multi-objective
 
 
 ## [1.2.2] bloa_stable